Feature: create and save new events
when pressing "a" a new Screen opens to create a new event. On filling
the Inputs, a new icalendar Event object is created and written to the
calendar.
Pretty barebones still, as you can not load multiple calendars and only
a few event variables can be set (e.g. atm repeating isn't implemented or
alarm)
NewEventScreen might need a slight rewrite to be more compact at some
point